[home] Add a typed API layer for the kernel + a fully defined mock kernel
This adds Kernel.ts, which is a typed interface for the kernel module and is its API layer. I also defined a platform layer for the kernel that runs on web as well as on iOS & Android when Home is run as a regular project instead of as the privileged Home app.